feat(eslint): disable perfectionist sort-object-types rule
Add a configuration to disable the 'perfectionist/sort-object-types' rule in the ESLint settings. This change aims to enhance flexibility in object type sorting within the ESLint configuration, allowing for a more tailored coding style.

Additionally, update the TypeScript import configuration to include paths in the 'typescriptAliases' object, improving the handling of TypeScript aliases and enhancing overall clarity in the configuration.